Only call flush once per instr
It looks like this was actually happening a few times per instr,
needlessly segmenting DBG_VALUEs, leading to different output orders.
Probably a suprious change, but lets swallow it in an apparently safe way
now so that we can compare object files in the future.

There's a tiny risk that we re-order identical DBG_VALUEs if a variable is
affected by an instruction twice. It seems unlikely*, but out of caution
use stable_sort for now. If this turns out to be expensive, profiling will
tell us that.

* edit: proven to happen now in copies...