Mangle a ton of "ldv-as-analysis" things
Basically: we can't handle both "modes" of variable location production at
the same time, because InstructionOrdering considers DBG_VALUEs following
an inst to be the same order number as that inst; while the history emitter
is more precise about where a location starts and stops. Plus, the
DBG_VALUE emitter is happy to create ranges bewteen two DBG_VALUE
instructions (or end-of-block instrs) that don't actually cover any normal
instructions.

Add a HAHA_... flag for switching behaviour between each mode. This sucks,
but it's necessary. Insert lots of hacks into the location list builder and
the label-producer, as the presence of DBG_VALUEs can change the line table
for moderately crazy reasons.

There are a few bits and pieces here to try and reduce the number of no-op
ranges produced. This could be even stronger, but it gets harder and harder
to prove equivalence.