Refactor stack-slot mashing code
In these code segments, we need to identify what the largest super-register
that fits in the stack slots would be, so that we can go and clobber or
transfer all the subregisters.

We're also now terminating the search once we find a superregister that
fills the whole slot -- this is because arm8.7a has some weird extensions
that let you load a whole cache-line into a group of registers, making the
largest super register 2^12 bits wide. With one meeeellliiiooonnn subregs.