Experiment: what if we do something simpler for simple assignments
i.e.: if we don't need to produce the PHI locations that join together
values from multiple assignments, because there's only one assignment,  how
much performance do we save? This isn't a real solution because we need to
have PHIs at the dominance frontier of a single dbg.value, but that's
easier to do.