Try just shoving a bit in ilist_iterator
This should give us an idea of what the memory and runtime costs are of having
a _separate_ bit hanging around, not folded into the node pointer as tried
before. Given how much iterators are promoted and passed-by-value, it might be
that the overhead is negligable.

Note that this uh, doesn't work, there are various bits of LLVM that break
because now they get fed debug-info where they're not expecting it. However
it's good enough to do the compile time checking, so let's do that.