ihold my beer,
The reason all these changes are needed in the first place are that
we're not funnelling DIArgList references through MAV/VAM pairings,
which eases some later DebugValueUser stuff. However: we don't need to
do that now, and could potentially fold it into a restructure and
optimisation later on when dbg.assign support arrives. And it's much
more important to not annoy people upstream in the meantime.