[DebugInfo][RemoveDIs] Clarify SROAs interactions with debug-intrinsics
SROA can slice-up code like this into individual vector operations:

  store <3 x i8> %val, ptr %retval, align 4
  %load = load i32, ptr %retval, align 4
  call void @llvm.dbg.value(metadata i32 %load,

And when it does, it inserts all the new instructions immediately in front
of the dbg.value intrinsic, which ensures that the reconstructed value is
available when the dbg.value "executes". With RemoveDIs non-instruction
debug-info, this doesn't happen by default as you can't usually insert "at"
the place where a debug-info record exists. To get around this, manually
set the "head" bit of the position iterator, which requests that newly
inserted instructions come before any DPValues attached to that position.

This is tested by function @PR14212 in vector-promotion.ll, where the
second dbg.value in the function will float upwards and come before the
rewritten instructions representing the load.