Fix gnochange issues -- scenarios where debug-info can change codegen
The vast majority of these issue are just swapping out std::next or an
iterator increment with getNextNonDebugInstruction. A few are counting
instructions; one or two are accidentally dropping the head bit by
dereferencing the iterator to an instruction pointer.

In a few scenarios we call getStableDebugLoc instead of getDebugLoc because
it's a site where we definitely pick a dbg.value's line number for a real
instruction, causing .debug_line to change between -g and -gmlt.

I'm not really planning on writing tests for all these changes as they're
numerous and diverse.

(Contains D151419)