Arguable pending head-bit situation
This modification seems correct -- we're skipping over debug intrinsics,
therefore a corresponding iterator would not have the "head" bit set.
However, I've so far failed to come up with a test case that this affects,
therefore I haven't tried to send it upstream. It's likely going to
languish here until we get a proper reproducer indicating it's necessary.