Turn off salvaging+sinking of debug-info for sunk instructions
We've got a non-production-ready implementation of this for DPValues,
however it fills me with sadness. The debug-info use-before-defs that this
code generates is something that instruction referencing in the codegen
pipeline is designed to handle, however we can't rely on that behaviour as
not all backends use instruction referencing. Yet.

This isn't a difficult piece of code to handle, we just don't have any
enthusiasm for getting it done ASAP.