Turn off loop-strength-reduction salvaging for comparison purposes
This piece of code has no relationship with the instruction API work we've
been performing, but it's also tightly coupled with the dbg.value storage
model. Turn it off for comparison purposes for the moment: we can easily
re-implement the code for DPValues, but would probably use templates, so
we've got little interest in doing that right now.