Fix a debug-info-affects-codegen problem in merge-icmps for comparisons
This is an interesting problem, see the comment added, because merge-icmps
splitting blocks on debug-info is a debug-affects-codegen issue but by
chance it usually just causes variable locations to be artifially
preserved, rather than for the output to change. Exactly what to do about
this needs more study (but in the worst case, it's a debug-affects-codegen
issue that wants fixing).