Reapply c23608b8d58, which was reverted in 157b62612
It turned out that this debug-info patch depended on another one that I
wasn't going to land immediately; now that it's in, reapply the original.

Switch to using presence of the CU as an indicator that we're building with
debug-info. There are too many tests without DISubprograms for this to work
it seems, and it's not a verifier requirement that all functions have
DISubprograms.

Original commit message and PR num follows:

[DebugInfo][RemoveDIs] Don't pointlessly scan funcs for debug-info (#79327)

The utility functions this patch modifies are part of cleanly
transitioning from a context where we use dbg.value intrinsics to one
where we use DPValue objects to record debug-info, and back again.
However, this is a waste of time in non-debug builds (i.e. no -g on the
command line). We still have to call the function on all blocks though
to set the IsNewDbgInfoFormat flag.

To reduce the overhead of this, test whether there's any debug-info in
the function by checking whether the function has a DISubprogram, and
pass a flag down to the utility functions indicating whether they can
skip the scan.

It feels a bit dumb to me now that we're scanning and setting a flag in
a load of blocks when we don't have to -- however it's been really
valuable during development for working out where spurious dbg.value
intrinsics leak into a RemoveDIs context. Happily we'll be able to just
delete this flag entirely when RemoveDIs lands and sticks, and the
conversion routines will eventually be pushed down into the debug-info
autoupgrade path.